server_getskilled-main/requirements.txt
flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7 
python-dotenv
openai
flask-cors 